Moved id_check to authorize_idcheck
