+++ lib/openmaize/plugs/authorize.ex
@@ -31,9 +31,7 @@ defmodule Openmaize.Authorize do
 
   """
 
-  import Plug.Conn
-  import Openmaize.Report
-  alias Openmaize.Config
+  import Openmaize.Authorize.Base
 
   @behaviour Plug
 
@@ -53,33 +51,5 @@ defmodule Openmaize.Authorize do
   defp run(conn, opts, data) do
     get_match(conn) |> permitted?(data) |> authorized?(conn, opts)
   end
-
-  defp permitted?({{0, _}, path}, nil) do
-    {:error, "You have to be logged in to view #{path}"}
-  end
-  defp permitted?(_, nil), do: {:ok, :nomatch}
-  defp permitted?({{0, match_len}, path}, %{role: role}) do
-    match = :binary.part(path, {0, match_len})
-    if role in Map.get(Config.protected, match) do
-      {:ok, path, match}
-    else
-      {:error, role, "You do not have permission to view #{path}"}
-    end
-  end
-  defp permitted?(_, _), do: {:ok, :nomatch}
-
-  defp get_match(%Plug.Conn{request_path: path}) do
-    {:binary.match(path, Map.keys(Config.protected)), path}
-  end
-
-  def authorized?(:ok, conn, _), do: conn
-  def authorized?({:ok, :nomatch}, conn, _), do: put_private(conn, :openmaize_skip, true)
-  def authorized?({:ok, path, match}, conn, _) do
-    put_private(conn, :openmaize_vars, %{path: path, match: match})
-  end
-  def authorized?({:error, message}, conn, {false, _}), do: send_error(conn, 401, message)
-  def authorized?({:error, message}, conn, _), do: handle_error(conn, message)
-  def authorized?({:error, _, message}, conn, {false, _}), do: send_error(conn, 403, message)
-  def authorized?({:error, role, message}, conn, _), do: handle_error(conn, role, message)
 
 end
