+++ mix.exs
@@ -7,7 +7,7 @@ defmodule Openmaize.Mixfile do
 
   def project do
     [app: :openmaize,
-      version: "0.6.8",
+      version: "0.7.0",
       elixir: "~> 1.0",
       name: "Openmaize",
       description: @description,
